modify CRUD finance

+++ src/http/finance/controller.js
@@ -1,27 +1,18 @@
 const {
   tambahFinance,
   tampilKeuangan,
-  cariFinanceById,
   ubahFinance,
   hapusFinance,
+  cariFinanceById,
 } = require("../finance/service.js");
 
 const { resSukses, resGagal } = require("../../payloads/payload.js");
 
 const getAllFinance = async (req, res) => {
   try {
-    const data = await tampilKeuangan();
+    const userId = req.user.id;
+    const data = await tampilKeuangan(userId);
     return resSukses(res, 200, "succes", "Data keuangan", data);
-  } catch (error) {
-    return resGagal(res, 500, "error", error.message);
-  }
-};
-
-const getFinanceById = async (req, res) => {
-  try {
-    const id = req.params.id;
-    const data = await cariFinanceById(id);
-    return resSukses(res, 200, "success", "Data berdasarkan ID", data);
   } catch (error) {
     return resGagal(res, 500, "error", error.message);
   }
@@ -30,16 +21,16 @@ const getFinanceById = async (req, res) => {
 const createFinance = async (req, res) => {
   try {
     const { type, category, amount, date, note } = req.body;
-    const user_id = req.user.id;
-
-    const finance = await tambahFinance({
-      user_id,
+    const userId = req.user.id;
+    const body = {
+      user_id: userId,
       type,
       category,
       amount,
       date,
       note,
-    });
+    };
+    const finance = await tambahFinance(body);
     return resSukses(res, 201, "success", "Data berhasil ditambahkan", finance);
   } catch (error) {
     return resGagal(res, 500, "error", error.message);
@@ -49,20 +40,28 @@ const createFinance = async (req, res) => {
 const updateFinance = async (req, res) => {
   try {
     const id = parseInt(req.params.id);
+    const user_id = req.user.id;
 
+    const dataFinance = await cariFinanceById(id);
+
+    if (dataFinance.user_id !== user_id) {
+      return resGagal(res, 403, "error", "Maaf, akses ditolak");
+    }
     const { type, category, amount, date, note } = req.body;
 
-    const user_id = req.user.id;
-    const body = {
-      user_id,
-      type,
-      category,
-      amount,
-      date,
-      note,
-    };
+    const updateData = {};
 
-    const data = await ubahFinance(id, body);
+    if (type && type.trim() !== "") updateData.type = type.toLowerCase().trim();
+    if (category && category.trim() !== "") updateData.category = category;
+    if (amount) updateData.amount = amount;
+    if (note !== undefined) updateData.note = note;
+
+    if (date && date.trim() !== "") {
+      const [day, month, year] = date.split("-");
+      updateData.date = `${year}-${month}-${day}`;
+    }
+
+    const data = await ubahFinance(id, updateData);
     return resSukses(res, 200, "success", "Data berhasil diubah", data);
   } catch (error) {
     return resGagal(res, 500, "error", error.message);
@@ -72,7 +71,15 @@ const updateFinance = async (req, res) => {
 const deleteFinance = async (req, res) => {
   try {
     const id = parseInt(req.params.id);
-    const data = await hapusFinance(id);
+    const user_id = req.user.id;
+
+    const dataFinance = await cariFinanceById(id);
+
+    if (dataFinance.user_id !== user_id) {
+      return resGagal(res, 403, "error", "Maaf, akses ditolak");
+    }
+
+    await hapusFinance(id);
     return resSukses(res, 201, "success", "Data berhasil dihapus");
   } catch (error) {
     return resGagal(res, 500, "error", error.message);
@@ -81,7 +88,7 @@ const deleteFinance = async (req, res) => {
 
 module.exports = {
   getAllFinance,
-  getFinanceById,
+
   createFinance,
   updateFinance,
   deleteFinance,
